requirements.txt
gspread==5.7.2
httpx==0.23.3
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
oauth2client==4.1.3
